Fix bugs found by running the full pipeline end-to-end
- 02_load_data.sql: a literal "/*" inside a comment (from a *.csv glob
  reference) was read as a nested comment open with no matching close,
  breaking the whole script; CODEPAGE isn't supported by BULK INSERT on
  Linux SQL Server; RowCount is a reserved word and can't be used bare
  as a column alias.
- 03_kpi_views.sql: SQL Server's SUM() rejects BIT columns directly;
  cast ChargeOffFlag to INT once in the base view.
- build_dataset.py: CreditScore was left as float64 (e.g. "663.0"),
  which BULK INSERT rejects for a SMALLINT destination column.

Verified: full pipeline (schema -> load -> views -> procs) runs clean
against a live SQL Server 2022 container with all 25,000 loans loaded
and every KPI view/procedure returning correct results.

+++ scripts/build_dataset.py
@@ -266,7 +266,7 @@ def build_members(unsecured: pd.DataFrame) -> pd.DataFrame:
     )
     members["MemberID"] = np.arange(1, len(members) + 1)
 
-    members["CreditScore"] = rank_rescale_linear(members["CreditScoreRaw"], *CREDIT_SCORE_BAND)
+    members["CreditScore"] = rank_rescale_linear(members["CreditScoreRaw"], *CREDIT_SCORE_BAND).astype(int)
     members["AnnualIncome"] = rank_rescale(members["AnnualIncomeRaw"], *ANNUAL_INCOME_BAND)
 
     raw_dti = (members["MonthlyDebtRaw"] * 12) / members["AnnualIncomeRaw"]
